+++ src/app/signin/page.tsx
@@ -1,6 +1,6 @@
-import { EmailLoginForm } from '@/app/temp/EmailLoginForm';
-import { LoginHeader } from '@/app/temp/LoginHeader';
-import { SocialLogin } from '@/app/temp/SocialLogin';
+import { EmailLoginForm } from '@/app/signin/EmailLoginForm';
+import { LoginHeader } from '@/app/signin/LoginHeader';
+import { SocialLogin } from '@/app/signin/SocialLogin';
 
 export default function LoginPage() {
   return (
